+++ src/utils/pageRegion/getAttributes.ts
@@ -13,7 +13,7 @@ import { NullableNodePos } from "../../types/node";
 import { getStateFromContext } from "../editor";
 import { doesDocHavePageNodes } from "../nodes/page/page";
 import { getHeaderFooterNodeType, isHeaderFooterNode } from "../nodes/headerFooter/headerFooter";
-import { isBodyNode } from "../nodes/body";
+import { isBodyNode } from "../nodes/body/body";
 import { getPageNodeByPageNum, handleOutOfRangePageNum, isPageNumInRange } from "../nodes/page/pageNumber";
 
 /**
